API/requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
click==6.7
demjson==2.2.4
Flask==1.0.2
fuzzywuzzy==0.16.0
idna==2.5
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.14.5
olefile==0.45.1
pandas==0.23.3
Pillow==4.2.1
pytesseract==0.1.7
python-dateutil==2.7.3
pytz==2018.5
requests==2.18.1
lxml==3.5.0
six==1.11.0
urllib3==1.21.1
user-agent==0.1.8
Werkzeug==0.14.1
